requirements.txt
alembic==1.18.4
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
attrs==26.1.0
backports.asyncio.runner==1.2.0
blinker==1.9.0
Bottleneck==1.6.0
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cryptography==46.0.7
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.135.3
filetype==1.2.0
fonttools==4.62.1
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.2
google-auth==2.49.1
google-cloud-aiplatform==1.146.0
google-cloud-bigquery==3.41.0
google-cloud-core==2.5.1
google-cloud-resource-manager==1.17.0
google-cloud-storage==3.10.1
google-cloud-vectorsearch==0.9.0
google-crc32c==1.8.0
google-genai==1.70.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
greenlet==3.3.2
grpc-google-iam-v1==0.14.4
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
iniconfig==2.3.0
ipykernel==7.2.0
ipython==8.39.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
langchain==1.2.15
langchain-core==1.2.27
langchain-google-genai==4.2.1
langchain-google-vertexai==3.2.2
langchain-openai==1.1.12
langchain-tests==1.1.5
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.26
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
narwhals==2.19.0
nest-asyncio==1.6.0
nltk==3.9.4
numexpr==2.14.1
numpy==1.26.4
openai==2.30.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==24.2
pandas==2.3.3
parso==0.8.6
pillow==10.4.0
platformdirs==4.9.4
plotly==6.6.0
pluggy==1.6.0
prompt_toolkit==3.0.52
proto-plus==1.27.2
protobuf==5.29.6
psutil==7.2.2
psycopg2-binary==2.9.11
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.20.0
pyparsing==3.3.2
pytest==9.0.3
pytest-asyncio==1.3.0
pytest-benchmark==5.2.3
pytest-codspeed==4.3.0
pytest-recording==0.13.4
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.24
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.49
stack-data==0.6.3
starlette==1.0.0
streamlit==1.39.1
syrupy==5.1.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
toml==0.10.2
tomli==2.4.1
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
validators==0.35.0
vcrpy==8.1.1
watchdog==5.0.3
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
wordcloud==1.9.6
wrapt==2.1.2
xxhash==3.6.0
zstandard==0.25.0
